requirements.txt
torch
torchvision

numpy
pillow
matplotlib
seaborn
scikit-learn